#1 Some configs refactoring
