+++ package.json
@@ -8,10 +8,10 @@
   },
   "scripts": {
     "start": "cross-env NODE_ENV=production babel-node server -r scripts/serverPreconfig.js",
-    "start:dev": "nodemon --watch server --watch webpack.dev.config.js --watch .env --exec 'cross-env NODE_ENV=development babel-node server -r scripts/serverPreconfig.js'",
+    "start:dev": "nodemon --watch server --watch webpack/dev.config.js --watch .env --exec 'cross-env NODE_ENV=development babel-node server -r scripts/serverPreconfig.js'",
     "prebuild": "yarn build:clean",
-    "build": "cross-env NODE_ENV=production webpack --config webpack.prod.config.js --color -p --progress --hide-modules --display-optimization-bailout",
-    "build:dev": "cross-env NODE_ENV=development webpack --config webpack.dev.config.js --color -p --progress --hide-modules",
+    "build": "cross-env NODE_ENV=production webpack --config webpack/prod.config.js --color -p --progress --hide-modules --display-optimization-bailout",
+    "build:dev": "cross-env NODE_ENV=development webpack --config webpack/dev.config.js --color -p --progress --hide-modules",
     "build:clean": "rimraf ./static/bundles",
     "test": "jest --config=jest.config.js",
     "test:coverage": "yarn test --coverage",
@@ -107,6 +107,7 @@
     "ejs": "^2.6.1",
     "emoji-mart": "^2.9.2",
     "express": "^4.16.4",
+    "file-system": "^2.2.2",
     "formik": "^1.3.2",
     "glob": "^7.1.3",
     "ignore-styles": "^5.0.1",
