+++ server/setup.js
@@ -2,7 +2,7 @@ function setup(app) {
   const isProd = process.env.NODE_ENV === 'production';
 
   if (!isProd) {
-    const webpackConfig = require('../webpack.dev.config');
+    const webpackConfig = require('../webpack/dev.config');
     const addDevMiddlewares = require('./middlewares/addDevMiddlewares');
     addDevMiddlewares(app, webpackConfig);
   }
