--- webpack-stats-production.json
@@ -1,8 +1,0 @@
-{
-  "assetsByChunkName": {
-    "commons": "commons.48adcc895a1875c2c591.bundle-production.js",
-    "application": "application.48adcc895a1875c2c591.bundle-production.js",
-    "home-page": "home-page.48adcc895a1875c2c591.bundle-production.js",
-    "main": "main.48adcc895a1875c2c591.bundle-production.js"
-  }
-}
