+++ webpack/dev.config.js
@@ -1,17 +1,13 @@
-const path = require('path');
 const webpack = require('webpack');
 
-module.exports = require('./webpack.base.config')({
+module.exports = require('./base.config')({
   mode: 'development',
 
   // Emit a source map for easier debugging
   devtool: 'eval-source-map',
 
   // Add hot reloading in development
-  entry: [
-    'webpack-hot-middleware/client?reload=true',
-    path.join(process.cwd(), './src/app.js'),
-  ],
+  entry: ['webpack-hot-middleware/client?reload=true', './src/app.js'],
 
   // Don't use hashes in dev mode for better performance
   output: {
