+++ webpack/stats-production.json
@@ -1,0 +1,8 @@
+{
+  "assetsByChunkName": {
+    "commons": "commons.4a38c699411176623c87.bundle-production.js",
+    "application": "application.4a38c699411176623c87.bundle-production.js",
+    "home-page": "home-page.4a38c699411176623c87.bundle-production.js",
+    "main": "main.4a38c699411176623c87.bundle-production.js"
+  }
+}
